refactor(PROC-1576): lift SeriesHandle and seeded_series fixtures into conftest
The original test_edit.py author left a TODO to do this once a second test
file needed the same boilerplate. The redaction tests in the next PR want
the same SeriesHandle + fresh_series + seeded_series setup, so lifting now.

No behavior change: test_edit.py imports SeriesHandle from conftest and
inherits the fixtures.
